Created new files for tables.
