+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/brev/maler/vedlegg/OpplysningerBruktIBeregningUT.kt
@@ -8,7 +8,7 @@ import no.nav.pensjon.brev.api.model.LetterMetadata
 import no.nav.pensjon.brev.api.model.vedlegg.OpplysningerBruktIBeregningUTDto
 import no.nav.pensjon.brev.api.model.vedlegg.OpplysningerBruktIBeregningUTDto.BarnetilleggGjeldende.Fellesbarn
 import no.nav.pensjon.brev.latex.LaTeXCompilerService
-import no.nav.pensjon.brev.maler.vedlegg.OpplysningerBruktIBeregningUT.vedleggOpplysningerBruktIBeregningUTBarnetillegg
+import no.nav.pensjon.brev.maler.vedlegg.OpplysningerBruktIBeregningUT.vedleggOpplysningerBruktIBeregningUT
 import no.nav.pensjon.brev.template.Language
 import no.nav.pensjon.brev.template.Letter
 import no.nav.pensjon.brev.template.dsl.createTemplate
@@ -56,7 +56,7 @@ class OpplysningerBruktIBeregningUTTest {
                 )
 
             includeAttachment(
-                vedleggOpplysningerBruktIBeregningUTBarnetillegg,
+                vedleggOpplysningerBruktIBeregningUT,
                 Fixtures.create(OpplysningerBruktIBeregningUTDto::class).expr()
                 )
         }
